Move IngestionJobRow to a separate source file.
